Revert "feat: use http client with curl"
