+++ .gitignore
@@ -6,5 +6,3 @@ composer.lock
 .phpunit.result.cache
 /build
 .vscode
-
-/coverage
