+++ composer.json
@@ -7,8 +7,7 @@
       "ext-curl": "*",
       "ext-json": "*",
       "ext-mbstring": "*",
-    "psr/http-client": "^1.0",
-    "psr/http-message": "^2.0"
+      "guzzlehttp/guzzle":"^7"
   },
     "require-dev": {
       "phpunit/phpunit": "^9",
