+++ src/PatpassComercio/Inscription.php
@@ -17,7 +17,7 @@ use Transbank\Utils\HttpClientRequestService;
 use Transbank\Utils\RequestServiceTrait;
 use Transbank\Contracts\RequestService;
 use Transbank\PatpassComercio\Options;
-use Transbank\Utils\Curl\Exceptions\CurlRequestException;
+use GuzzleHttp\Exception\GuzzleException;
 
 class Inscription
 {
@@ -63,7 +63,7 @@ class Inscription
      * @param string $city
      *
      * @throws InscriptionStartException
-     * @throws CurlRequestException
+     * @throws GuzzleException
      *
      * @return InscriptionStartResponse
      */
@@ -123,7 +123,7 @@ class Inscription
      * @param string $token
      *
      * @throws InscriptionStatusException
-     * @throws CurlRequestException
+     * @throws GuzzleException
      *
      * @return InscriptionStatusResponse
      */
