--- src/Utils/Curl/Uri.php
@@ -1,167 +1,0 @@
-<?php
-
-namespace Transbank\Utils\Curl;
-
-use Psr\Http\Message\UriInterface;
-
-class Uri implements UriInterface
-{
-    private string $scheme = '';
-    private string $userInfo = '';
-    private string $host = '';
-    private int|null $port;
-    private string $path = '';
-    private string $query = '';
-    private string $fragment = '';
-
-    public function __construct(string $uri = '')
-    {
-        if ($uri) {
-            $this->parseUri($uri);
-        }
-    }
-
-    private function parseUri(string $uri): void
-    {
-        $parts = parse_url($uri);
-        if ($parts === false) {
-            throw new \InvalidArgumentException("Invalid URI: $uri");
-        }
-
-        $this->scheme = $parts['scheme'] ?? '';
-        $this->userInfo = isset($parts['user']) ? $parts['user'] : '';
-        if (isset($parts['pass'])) {
-            $this->userInfo .= ':' . $parts['pass'];
-        }
-        $this->host = $parts['host'] ?? '';
-        $this->port = $parts['port'] ?? null;
-        $this->path = $parts['path'] ?? '';
-        $this->query = $parts['query'] ?? '';
-        $this->fragment = $parts['fragment'] ?? '';
-    }
-
-    public function getScheme(): string
-    {
-        return $this->scheme;
-    }
-
-    public function withScheme($scheme): UriInterface
-    {
-        $new = clone $this;
-        $new->scheme = $scheme;
-        return $new;
-    }
-
-    public function getAuthority(): string
-    {
-        $authority = $this->host;
-        if ($this->userInfo) {
-            $authority = $this->userInfo . '@' . $authority;
-        }
-        if ($this->port !== null) {
-            $authority .= ':' . $this->port;
-        }
-        return $authority;
-    }
-
-    public function getUserInfo(): string
-    {
-        return $this->userInfo;
-    }
-
-    public function withUserInfo($user, $password = null): UriInterface
-    {
-        $new = clone $this;
-        $new->userInfo = $user;
-        if ($password !== null) {
-            $new->userInfo .= ':' . $password;
-        }
-        return $new;
-    }
-
-    public function getHost(): string
-    {
-        return $this->host;
-    }
-
-    public function withHost($host): UriInterface
-    {
-        $new = clone $this;
-        $new->host = $host;
-        return $new;
-    }
-
-    public function getPort(): ?int
-    {
-        return $this->port;
-    }
-
-    public function withPort($port): UriInterface
-    {
-        $new = clone $this;
-        $new->port = $port;
-        return $new;
-    }
-
-    public function getPath(): string
-    {
-        return $this->path;
-    }
-
-    public function withPath($path): UriInterface
-    {
-        $new = clone $this;
-        $new->path = $path;
-        return $new;
-    }
-
-    public function getQuery(): string
-    {
-        return $this->query;
-    }
-
-    public function withQuery($query): UriInterface
-    {
-        $new = clone $this;
-        $new->query = $query;
-        return $new;
-    }
-
-    public function getFragment(): string
-    {
-        return $this->fragment;
-    }
-
-    public function withFragment($fragment): UriInterface
-    {
-        $new = clone $this;
-        $new->fragment = $fragment;
-        return $new;
-    }
-
-    public function __toString(): string
-    {
-        $uri = '';
-
-        if ($this->scheme) {
-            $uri .= $this->scheme . '://';
-        }
-
-        $authority = $this->getAuthority();
-        if ($authority) {
-            $uri .= $authority;
-        }
-
-        $uri .= $this->path;
-
-        if ($this->query) {
-            $uri .= '?' . $this->query;
-        }
-
-        if ($this->fragment) {
-            $uri .= '#' . $this->fragment;
-        }
-
-        return $uri;
-    }
-}
