+++ src/Utils/HttpClientRequestService.php
@@ -5,10 +5,9 @@ namespace Transbank\Utils;
 use Transbank\Contracts\HttpClientInterface;
 use Transbank\Contracts\RequestService;
 use Transbank\Utils\TransbankApiRequest;
-use Transbank\Utils\Curl\HttpCurlClient;
 use Transbank\Webpay\Exceptions\WebpayRequestException;
 use Transbank\Webpay\Options;
-use Transbank\Utils\Curl\Exceptions\CurlRequestException;
+use GuzzleHttp\Exception\GuzzleException;
 use Psr\Http\Message\ResponseInterface;
 
 class HttpClientRequestService implements RequestService
@@ -30,7 +29,7 @@ class HttpClientRequestService implements RequestService
 
     public function __construct(HttpClientInterface|null $httpClient = null)
     {
-        $this->setHttpClient($httpClient ?? new HttpCurlClient());
+        $this->setHttpClient($httpClient ?? new HttpClient());
     }
 
     /**
@@ -57,7 +56,7 @@ class HttpClientRequestService implements RequestService
      * @param array   $payload
      * @param Options $options
      *
-     * @throws CurlRequestException
+     * @throws GuzzleException
      * @throws WebpayRequestException
      *
      * @return array
