+++ src/Webpay/Oneclick/MallTransaction.php
@@ -12,7 +12,7 @@ use Transbank\Webpay\Oneclick\Responses\MallTransactionAuthorizeResponse;
 use Transbank\Webpay\Oneclick\Responses\MallTransactionCaptureResponse;
 use Transbank\Webpay\Oneclick\Responses\MallTransactionRefundResponse;
 use Transbank\Webpay\Oneclick\Responses\MallTransactionStatusResponse;
-use Transbank\Utils\Curl\Exceptions\CurlRequestException;
+use GuzzleHttp\Exception\GuzzleException;
 
 class MallTransaction
 {
@@ -31,7 +31,7 @@ class MallTransaction
      * @return MallTransactionAuthorizeResponse
      *
      * @throws MallTransactionAuthorizeException
-     * @throws CurlRequestException
+     * @throws GuzzleException
      */
     public function authorize(
         string $userName,
@@ -74,7 +74,7 @@ class MallTransaction
      * @return MallTransactionCaptureResponse
      *
      * @throws MallTransactionCaptureException
-     * @throws CurlRequestException
+     * @throws GuzzleException
      */
     public function capture(
         string $childCommerceCode,
@@ -114,7 +114,7 @@ class MallTransaction
      * @return MallTransactionStatusResponse
      *
      * @throws MallTransactionStatusException
-     * @throws CurlRequestException
+     * @throws GuzzleException
      */
     public function status(string $buyOrder): MallTransactionStatusResponse
     {
@@ -146,7 +146,7 @@ class MallTransaction
      * @return MallTransactionRefundResponse
      *
      * @throws MallRefundTransactionException
-     * @throws CurlRequestException
+     * @throws GuzzleException
      */
     public function refund(
         string $buyOrder,
