+++ src/Webpay/TransaccionCompleta/MallTransaction.php
@@ -21,7 +21,7 @@ use Transbank\Webpay\TransaccionCompleta\Responses\MallTransactionRefundResponse
 use Transbank\Webpay\TransaccionCompleta\Responses\MallTransactionStatusResponse;
 use Transbank\Webpay\TransaccionCompleta\Responses\MallTransactionCaptureResponse;
 use Transbank\Utils\InteractsWithWebpayApi;
-use Transbank\Utils\Curl\Exceptions\CurlRequestException;
+use GuzzleHttp\Exception\GuzzleException;
 use Transbank\Webpay\Exceptions\WebpayRequestException;
 
 class MallTransaction
@@ -224,7 +224,7 @@ class MallTransaction
      * @param int|float  $captureAmount
      *
      * @throws MallTransactionCaptureException
-     * @throws CurlRequestException
+     * @throws GuzzleException
      *
      * @return MallTransactionCaptureResponse
      */
