+++ tests/Webpay/WebpayPlus/WebpayPlusWithoutMocksTest.php
@@ -10,8 +10,6 @@ use Transbank\Webpay\WebpayPlus\Exceptions\TransactionCommitException;
 use Transbank\Webpay\WebpayPlus\Exceptions\TransactionCreateException;
 use Transbank\Webpay\WebpayPlus\Transaction;
 use Transbank\Webpay\WebpayPlus\Responses\TransactionStatusResponse;
-use Transbank\Utils\Curl\HttpCurlClient;
-use Transbank\Utils\HttpClientRequestService;
 
 class WebpayPlusWithoutMocksTest extends TestCase
 {
@@ -41,8 +39,6 @@ class WebpayPlusWithoutMocksTest extends TestCase
      */
     public $options;
 
-    public $clientRequestService;
-
     protected function setUp(): void
     {
         $this->amount = 1000;
@@ -55,14 +51,12 @@ class WebpayPlusWithoutMocksTest extends TestCase
             WebpayPlus::INTEGRATION_COMMERCE_CODE,
             Options::ENVIRONMENT_INTEGRATION
         );
-        $httpClient = new HttpCurlClient();
-        $this->clientRequestService = new HttpClientRequestService($httpClient);
     }
 
     /** @test */
     public function it_creates_a_real_transaction_with_options()
     {
-        $transaction = (new Transaction($this->options, $this->clientRequestService));
+        $transaction = (new Transaction($this->options));
         $transactionResult = $transaction->create(
             $this->buyOrder,
             $this->sessionId,
@@ -80,7 +74,7 @@ class WebpayPlusWithoutMocksTest extends TestCase
 
         $this->expectException(TransactionCreateException::class);
         $this->expectExceptionMessage('Not Authorized');
-        $transaction = (new Transaction($options, $this->clientRequestService));
+        $transaction = (new Transaction($options));
         $transaction->create($this->buyOrder, $this->sessionId, $this->amount, $this->returnUrl);
     }
 
@@ -113,7 +107,7 @@ class WebpayPlusWithoutMocksTest extends TestCase
     /** @test */
     public function it_can_not_commit_a_recently_created_transaction()
     {
-        $response = (new Transaction($this->options, $this->clientRequestService))->create(
+        $response = (new Transaction($this->options))->create(
             $this->buyOrder,
             $this->sessionId,
             $this->amount,
@@ -133,7 +127,7 @@ class WebpayPlusWithoutMocksTest extends TestCase
             WebpayPlus::INTEGRATION_DEFERRED_COMMERCE_CODE,
             Options::ENVIRONMENT_INTEGRATION
         );
-        $response = (new Transaction($deferredOptions, $this->clientRequestService))->create(
+        $response = (new Transaction($deferredOptions))->create(
             $this->buyOrder,
             $this->sessionId,
             $this->amount,
@@ -151,7 +145,7 @@ class WebpayPlusWithoutMocksTest extends TestCase
             WebpayPlus::INTEGRATION_API_KEY,
             WebpayPlus::INTEGRATION_COMMERCE_CODE,
             Options::ENVIRONMENT_INTEGRATION
-        ), $this->clientRequestService);
+        ));
         $response = $transaction->create($this->buyOrder, $this->sessionId, $this->amount, $this->returnUrl);
         $this->expectException(TransactionCaptureException::class);
         $this->expectExceptionMessage('Operation not allowed');
@@ -163,7 +157,7 @@ class WebpayPlusWithoutMocksTest extends TestCase
     /** @test */
     public function it_returns_a_card_number_in_null_when_it_not_exists()
     {
-        $transaction = new Transaction($this->options, $this->clientRequestService);
+        $transaction = new Transaction($this->options);
         $createResponse = $transaction->create($this->buyOrder, $this->sessionId, $this->amount, $this->returnUrl);
         $statusResponse = $transaction->status($createResponse->getToken());
